+++ apps/relay/src/channels.ts
@@ -1,59 +1,65 @@
-import { Redis } from "ioredis";
-import { ResultAsync, err, ok } from "neverthrow";
 import { randomUUID } from "crypto";
-import { RelayAsyncResult, RelayError } from "./errors";
+import { Hono } from "hono";
+import { generateNonce } from "siwe";
+import { z } from "zod";
+import { rateLimiter } from "hono-rate-limiter";
+import { channels as channelsMiddleware } from "./middlewares/channels";
+import { getConfig } from "./utils/getConfig";
+import { RelayError } from "./utils/errors";
+import { fromZodError } from "zod-validation-error";
 
-interface ChannelStoreOpts {
-  redisUrl: string;
-  ttl?: number;
+const { baseUrl } = getConfig();
+
+export const channels = new Hono().use(
+  rateLimiter({
+    limit: 1000,
+  }),
+  channelsMiddleware,
+);
+
+function constructUrl(parameters: { channelToken: string; nonce: string } & z.infer<typeof createSchema>): string {
+  const query = new URLSearchParams(parameters);
+  return `${baseUrl}?${query.toString()}`;
 }
 
-export class ChannelStore<T> {
-  private redis: Redis;
-  private ttl: number;
+const createSchema = z.object({
+  siweUri: z.string().url(),
+  domain: z.string(),
+  nonce: z.string().optional(),
+  notBefore: z.string().optional(),
+  expirationTime: z.string().optional(),
+  requestId: z.string().optional(),
+});
 
-  constructor({ redisUrl, ttl }: ChannelStoreOpts) {
-    this.redis = new Redis(redisUrl);
-    this.ttl = ttl ?? 3600;
+channels.post("/create", async (c) => {
+  const parseResult = createSchema.safeParse(c.req.json());
+  if (!parseResult.success) {
+    c.status(400);
+    return c.json({ error: "Validation error", message: fromZodError(parseResult.error) });
   }
+  const body = parseResult.data;
 
-  async open(state?: T): RelayAsyncResult<string> {
-    const channelToken = randomUUID();
-    return ResultAsync.fromPromise(
-      this.redis.set(channelToken, JSON.stringify(state ?? {}), "EX", this.ttl),
-      (err) => new RelayError("unavailable", err as Error),
-    ).andThen(() => ok(channelToken));
-  }
+  const channelToken = randomUUID();
+  try {
+    await c.var.channels.create(channelToken);
 
-  async update(channelToken: string, state: T): RelayAsyncResult<T> {
-    return ResultAsync.fromPromise(
-      this.redis.set(channelToken, JSON.stringify(state), "KEEPTTL"),
-      (err) => new RelayError("unavailable", err as Error),
-    ).andThen(() => ok(state));
-  }
+    const nonce = body.nonce ?? generateNonce();
+    const connectUri = constructUrl({ channelToken, nonce, ...body });
 
-  async read(channelToken: string): RelayAsyncResult<T> {
-    return ResultAsync.fromPromise(
-      this.redis.get(channelToken),
-      (err) => new RelayError("unavailable", err as Error),
-    ).andThen((channel) => {
-      if (channel) {
-        return ok(JSON.parse(channel));
-      } else {
-        return err(new RelayError("not_found", "Channel not found"));
-      }
-    });
-  }
+    const pendingChannel = {
+      state: "pending",
+      nonce,
+      url: connectUri,
+      connectUri,
+    } as const;
 
-  async close(channelToken: string) {
-    return ResultAsync.fromPromise(this.redis.del(channelToken), (err) => new RelayError("unknown", err as Error));
-  }
+    await c.var.channels.update(channelToken, pendingChannel);
+    c.status(201);
+    return c.json(pendingChannel);
+  } catch (e) {
+    if (!(e instanceof RelayError)) throw new Error("Unexpected error");
 
-  async clear() {
-    return this.redis.flushall();
+    c.status(500);
+    return c.json({ error: e.message });
   }
-
-  async stop() {
-    return this.redis.quit();
-  }
-}
+});