+++ apps/relay/src/utils/getCustodyAddress.ts
@@ -1,0 +1,13 @@
+import { getConfig } from "./getConfig";
+import { ID_REGISTRY_ADDRESS, idRegistryABI } from "@farcaster/core";
+
+const { publicClient } = getConfig();
+
+export async function getCustodyAddress(fid: number) {
+  return publicClient.readContract({
+    address: ID_REGISTRY_ADDRESS,
+    abi: idRegistryABI,
+    functionName: "custodyOf",
+    args: [BigInt(fid ?? 0)],
+  });
+}