+++ examples/with-next-auth/pages/api/auth/[...nextauth].ts
@@ -2,6 +2,7 @@ import NextAuth from "next-auth";
 import CredentialsProvider from "next-auth/providers/credentials";
 import { createAppClient, viemConnector } from "@farcaster/auth-client";
 import { NextApiRequest, NextApiResponse } from "next";
+import { JsonRpcProvider } from "ethers";
 
 export default (req: NextApiRequest, res: NextApiResponse) =>
   NextAuth(req, res, {