+++ packages/core/src/clients/ethereum/connector.ts
@@ -1,7 +1,5 @@
 import { type Hex } from "viem";
-import type { Provider } from "ethers";
 
 export interface EthereumConnector {
   getFid: (custody: Hex) => Promise<bigint>;
-  provider: Provider;
 }