+++ packages/core/src/clients/ethereum/viemConnector.ts
@@ -2,7 +2,6 @@ import { type Hex, createPublicClient, http } from "viem";
 import { optimism } from "viem/chains";
 import { ID_REGISTRY_ADDRESS, idRegistryABI } from "../../contracts/idRegistry.js";
 import { type EthereumConnector } from "./connector.js";
-import { JsonRpcProvider } from "ethers";
 
 interface ViemConfigArgs {
   rpcUrl?: string;
@@ -23,18 +22,7 @@ export const viemConnector = (args?: ViemConfigArgs): EthereumConnector => {
     });
   };
 
-  const getProvider = () => {
-    const { chain, transport } = publicClient;
-    const network = {
-      chainId: chain.id,
-      name: chain.name,
-    };
-    const rpc = transport.url ?? chain.rpcUrls.default.http[0];
-    return new JsonRpcProvider(rpc, network);
-  };
-
   return {
     getFid,
-    provider: getProvider(),
   };
 };